refactor(config): extract shared config helpers
